requirements.txt
vnpy_ctptest
vnpy_ctp
python-dotenv
flask==2.2.5
flask_jsonrpc==2.2.2
kafka-python==2.0.2